jolybell_clone_backend-dev-eldos/pyproject.toml
[tool.poetry]
name = "jolybell_backend"
version = "0.1.0"
description = ""
authors = ["Eldos <eldos.baktybekov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
environs = "^9.5.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
